code/requirements.txt
albumentations==1.2.1
imgaug==0.4.0
ml_collections==0.1.1
numpy==1.20.3
opencv-python==4.6.0.66
Pillow==8.2.0
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
SimpleITK==2.1.1.2
six==1.16.0
tensorboard==2.4.1
tensorboardX==2.4
torch==1.13.1
torchvision==0.14.1
tqdm==4.62.3